Fix up tasks
