+++ posthog/tasks/__init__.py
@@ -5,14 +5,17 @@ from . import (
     calculate_cohort,
     check_clickhouse_schema_drift,
     demo_create_data,
+    demo_reset_master_team,
     email,
     exporter,
     process_scheduled_changes,
+    prompts,
     split_person,
     sync_all_organization_available_features,
     tasks,
     usage_report,
     user_identify,
+    verify_persons_data_in_sync,
     warehouse,
 )
 
@@ -21,13 +24,16 @@ __all__ = [
     "calculate_cohort",
     "check_clickhouse_schema_drift",
     "demo_create_data",
+    "demo_reset_master_team",
     "email",
     "exporter",
     "process_scheduled_changes",
+    "prompts",
     "split_person",
     "sync_all_organization_available_features",
     "tasks",
-    "user_identify",
     "usage_report",
+    "user_identify",
+    "verify_persons_data_in_sync",
     "warehouse",
 ]
